@@ -436,10 +436,30 @@ class AttributeResolver
             }
             $props = array_map('trim', explode(',', trim($props)));
             $columns = [];
+            $xFkColumnNames = [];
+            foreach ($this->attributes as $key => $value) {
+                if (!empty($value->fkColName)) {
+                    $xFkColumnNames[$value->fkColName] = $key;
+                }
+            }
             foreach ($props as $prop) {
+                // for more info see test tests/specs/fk_col_name/fk_col_name.yaml
+                // File: ForeignKeyColumnNameTest::testIndexForColumnWithCustomName
+                // first check direct column names
                 if (!isset($this->attributes[$prop])) {
-                    throw new InvalidDefinitionException('Invalid index definition - property ' . $prop
-                        . ' not declared');
+                    // then check x-fk-column-name
+                    if (!in_array($prop, array_keys($xFkColumnNames))) {
+                        // then check relations/reference e.g. `user`/`user_id`
+                        $refPropName = (substr($prop, -3) === '_id') ? rtrim($prop, '_id') : null;
+                        if ($refPropName && !isset($this->attributes[$refPropName])) {
+                            throw new InvalidDefinitionException('Invalid index definition - property ' . $prop
+                                . ' not declared');
+                        } else {
+                            $prop = $refPropName;
+                        }
+                    } else {
+                        $prop = $xFkColumnNames[$prop];
+                    }
                 }
                 $columns[] = $this->attributes[$prop]->columnName;
             }
