@@ -12,7 +12,7 @@ class m200000_000005_change_table_v2_comments extends \yii\db\Migration
         $this->addColumn('{{%v2_comments}}', 'user_id', $this->bigInteger()->null()->defaultValue(null)->after('post_id'));
         $this->dropColumn('{{%v2_comments}}', 'author_id');
         $this->alterColumn('{{%v2_comments}}', 'message', $this->text()->notNull());
-        $this->alterColumn('{{%v2_comments}}', 'meta_data', $this->string(300)->null()->defaultValue(""));
+        $this->alterColumn('{{%v2_comments}}', 'meta_data', $this->string(300)->null()->defaultValue(''));
         $this->alterColumn('{{%v2_comments}}', 'created_at', $this->timestamp()->notNull());
         $this->addForeignKey('fk_v2_comments_post_id_v2_posts_id', '{{%v2_comments}}', 'post_id', '{{%v2_posts}}', 'id');
         $this->addForeignKey('fk_v2_comments_user_id_v2_users_id', '{{%v2_comments}}', 'user_id', '{{%v2_users}}', 'id');
