@@ -12,6 +12,7 @@ namespace app\models\base;
  * @property array $json3
  * @property array $json4
  * @property string $status
+ * @property string $status_x
  * @property string $search
  *
  */
@@ -25,7 +26,7 @@ abstract class Custom extends \yii\db\ActiveRecord
     public function rules()
     {
         return [
-            'trim' => [['status'], 'trim'],
+            'trim' => [['status', 'status_x'], 'trim'],
             'num_integer' => [['num'], 'integer'],
             'num_default' => [['num'], 'default', 'value' => 0],
             'json1_default' => [['json1'], 'default', 'value' => []],
@@ -48,6 +49,12 @@ abstract class Custom extends \yii\db\ActiveRecord
                 'draft',
             ]],
             'status_default' => [['status'], 'default', 'value' => 'draft'],
+            'status_x_string' => [['status_x'], 'string', 'max' => 10],
+            'status_x_in' => [['status_x'], 'in', 'range' => [
+                'active',
+                'draft',
+            ]],
+            'status_x_default' => [['status_x'], 'default', 'value' => 'draft'],
             'safe' => [['json1', 'json2', 'json3', 'json4'], 'safe'],
         ];
     }
