@@ -9,8 +9,8 @@ class m200000_000000_create_table_fruits extends \yii\db\Migration
     {
         $this->createTable('{{%fruits}}', [
             'ts' => $this->timestamp()->null()->defaultExpression("(CURRENT_TIMESTAMP)"),
-            'ts2' => $this->timestamp()->null()->defaultValue("2011-11-11 00:00:00"),
-            'ts3' => $this->timestamp()->null()->defaultValue("2022-11-11 00:00:00"),
+            'ts2' => $this->timestamp()->null()->defaultValue('2011-11-11 00:00:00'),
+            'ts3' => $this->timestamp()->null()->defaultValue('2022-11-11 00:00:00'),
             0 => '"ts4" timestamp NULL DEFAULT \'2022-11-11 00:00:00\'',
             1 => '"ts5" timestamp NULL DEFAULT (CURRENT_TIMESTAMP)',
             2 => '"ts6" timestamp NULL DEFAULT \'2000-11-11 00:00:00\'',
